rlichess: R Client and Analysis Tools for Lichess

Provides an interface to the Lichess API (https://lichess.org/api) for downloading chess game data, player profiles, and ratings, along with tidy tools for analyzing chess performances, openings, and trends.

See also

Author

Maintainer: Hiroki Yagi hiro.yagi0245@gmail.com

Authors:

Hiroki Yagi hiro.yagi0245@gmail.com